services/communication-ai/requirements.txt
fastapi
uvicorn[standard]
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
python-dotenv
